refactor errors classes
