+++ lib/whatsapp_sdk/resource/error.rb
@@ -4,6 +4,19 @@
 module WhatsappSdk
   module Resource
     module Error
+      class MissingArgumentError < StandardError
+        extend T::Sig
+
+        sig { returns(String) }
+        attr_reader :message
+
+        sig { params(message: String).void }
+        def initialize(message)
+          @message = message
+          super(message)
+        end
+      end
+
       class MissingValue < WhatsappSdk::Error
         extend T::Sig
